+++ src/senaite/astm/adapters/biomerieux/__init__.py
@@ -1,3 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from . import mini_vidas  # noqa: F401