+++ src/senaite/astm/instruments/dca_vantage.py
@@ -1,6 +1,10 @@
 # -*- coding: utf-8 -*-
 
+import re
+
 from senaite.astm import records
+from senaite.astm.core.instrument import Instrument
+from senaite.astm.core.instrument import register_instrument
 from senaite.astm.fields import ComponentField
 from senaite.astm.fields import ConstantField
 from senaite.astm.fields import DateTimeField
@@ -13,7 +17,7 @@ VERSION = "1.0.0"
 
 # Siemens Healthcare Diagnostics, DCA Vantage® Analyzer, Host Computer
 # Communications Link, 17306 Rev. E 2012-06, 2012
-HEADER_RX = r".*(DCA VANTAGE|DCA Vantage)\^"
+HEADER_RX = re.compile(rb".*(DCA VANTAGE|DCA Vantage)\^")
 
 PROCESSING_IDS = (
     "P",  # P: Production (the message contains clinical results)
@@ -40,32 +44,6 @@ RESULT_STATUSES = (
     "F",  # F: Final Result
     "C",  # C: Correction of previously transmitted results
 )
-
-
-def get_metadata(wrapper):
-    """Additional metadata
-    :param wrapper: The wrapper instance
-    :returns: dictionary of additional metadata
-    """
-    return {
-        "version": VERSION,
-        "header_rx": HEADER_RX,
-    }
-
-
-def get_mapping():
-    """Returns the wrappers for this instrument
-    """
-    return {
-        "H": HeaderRecord,
-        "P": PatientRecord,
-        "O": OrderRecord,
-        "R": ResultRecord,
-        "C": CommentRecord,
-        "Q": RequestInformationRecord,
-        "M": ManufacturerInfoRecord,
-        "L": TerminatorRecord,
-    }
 
 
 class HeaderRecord(records.HeaderRecord):
@@ -178,3 +156,25 @@ class ManufacturerInfoRecord(records.ManufacturerInfoRecord):
 class TerminatorRecord(records.TerminatorRecord):
     """Message Termination Record (L)
     """
+
+
+@register_instrument
+class DCAVantage(Instrument):
+    name = "dca_vantage"
+    header_regex = HEADER_RX
+    record_map = {
+        "H": HeaderRecord,
+        "P": PatientRecord,
+        "O": OrderRecord,
+        "R": ResultRecord,
+        "C": CommentRecord,
+        "Q": RequestInformationRecord,
+        "M": ManufacturerInfoRecord,
+        "L": TerminatorRecord,
+    }
+
+    def get_metadata(self, wrapper):
+        return {"version": VERSION, "header_rx": HEADER_RX.pattern.decode()}
+
+
+INSTRUMENT = DCAVantage()