+++ src/senaite/astm/tests/test_genexpert.py
@@ -28,7 +28,7 @@ class GeneXpert(ASTMTestBase):
         # Mock transport and protocol objects
         self.transport = self.get_mock_transport()
         self.protocol.transport = self.transport
-        self.mapping = genexpert.get_mapping()
+        self.mapping = genexpert.INSTRUMENT.record_map
 
     def get_mock_transport(self, ip="127.0.0.1", port=12345):
         transport = MagicMock()